requirements.txt
numpy==1.26.4
matplotlib==3.8.4
h5py==3.11.0
Pillow==10.4.0
tqdm==4.66.4
PyYAML>=6.0
